Version 2.18d
Version 2.18d

+++ src/main/java/nc/multiblock/turbine/tile/TileTurbineDynamoCoil.java
@@ -108,7 +108,7 @@ public class TileTurbineDynamoCoil extends TileTurbinePartBase implements ITileE
 		//getWorld().setBlockState(getPos(), getWorld().getBlockState(getPos()), 2);
 	}
 	
-	public double contributeConductivity(short dynamoCoilCheckCount) {
+	public double contributeConductivity(int dynamoCoilCheckCount) {
 		EnumFacing flowDir = getMultiblock().flowDir;
 		
 		if (!isMultiblockAssembled() || flowDir == null) {
